+++ src/animus/database/sqlalchemy/models/library/__init__.py
@@ -1,0 +1,3 @@
+from .folder_model import FolderModel
+
+__all__ = ['FolderModel']